# Optimize WeakReferenceMessenger