--- a/CommunityToolkit.Mvvm/Messaging/Internals/System/Collections.Generic/Dictionary2.cs
+++ b/CommunityToolkit.Mvvm/Messaging/Internals/System/Collections.Generic/Dictionary2.cs
@@ -1,11 +1,11 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
 
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Diagnostics.Contracts;
 using System.Runtime.CompilerServices;
-using Microsoft.Collections.Extensions;
 
 namespace System.Collections.Generic;
 
@@ -15,7 +15,7 @@ namespace System.Collections.Generic;
 /// <typeparam name="TKey">The type of keys in the dictionary.</typeparam>
 /// <typeparam name="TValue">The type of values in the dictionary.</typeparam>
 [DebuggerDisplay("Count = {Count}")]
-internal class Dictionary2<TKey, TValue> : IDictionarySlim<TKey, TValue>
+internal class Dictionary2<TKey, TValue> : IDictionary2<TKey, TValue>
     where TKey : IEquatable<TKey>
     where TValue : class
 {
@@ -152,31 +152,31 @@ internal class Dictionary2<TKey, TValue> : IDictionarySlim<TKey, TValue>
         {
             ref Entry entry = ref entries[i];
 
-            if (entry.hashCode == hashCode && entry.key.Equals(key))
+            if (entry.HashCode == hashCode && entry.Key.Equals(key))
             {
                 if (last < 0)
                 {
-                    bucket = entry.next + 1;
+                    bucket = entry.Next + 1;
                 }
                 else
                 {
-                    entries[last].next = entry.next;
+                    entries[last].Next = entry.Next;
                 }
 
-                entry.next = StartOfFreeList - this.freeList;
+                entry.Next = StartOfFreeList - this.freeList;
 
 #if NETSTANDARD2_1 || NET6_0_OR_GREATER
                 if (RuntimeHelpers.IsReferenceOrContainsReferences<TKey>())
 #endif
                 {
-                    entry.key = default!;
+                    entry.Key = default!;
                 }
 
 #if NETSTANDARD2_1 || NET6_0_OR_GREATER
                 if (RuntimeHelpers.IsReferenceOrContainsReferences<TValue>())
 #endif
                 {
-                    entry.value = default!;
+                    entry.Value = default!;
                 }
 
                 this.freeList = i;
@@ -186,7 +186,7 @@ internal class Dictionary2<TKey, TValue> : IDictionarySlim<TKey, TValue>
             }
 
             last = i;
-            i = entry.next;
+            i = entry.Next;
         }
 
         return false;
@@ -213,12 +213,12 @@ internal class Dictionary2<TKey, TValue> : IDictionarySlim<TKey, TValue>
                 break;
             }
 
-            if (entries[i].hashCode == hashCode && entries[i].key.Equals(key))
+            if (entries[i].HashCode == hashCode && entries[i].Key.Equals(key))
             {
-                return ref entries[i].value!;
+                return ref entries[i].Value!;
             }
 
-            i = entries[i].next;
+            i = entries[i].Next;
         }
 
         int index;
@@ -227,7 +227,7 @@ internal class Dictionary2<TKey, TValue> : IDictionarySlim<TKey, TValue>
         {
             index = this.freeList;
 
-            this.freeList = StartOfFreeList - entries[this.freeList].next;
+            this.freeList = StartOfFreeList - entries[this.freeList].Next;
             this.freeCount--;
         }
         else
@@ -249,13 +249,13 @@ internal class Dictionary2<TKey, TValue> : IDictionarySlim<TKey, TValue>
 
         ref Entry entry = ref entries![index];
 
-        entry.hashCode = hashCode;
-        entry.next = bucket - 1;
-        entry.key = key;
-        entry.value = default!;
+        entry.HashCode = hashCode;
+        entry.Next = bucket - 1;
+        entry.Key = key;
+        entry.Value = default!;
         bucket = index + 1;
 
-        return ref entry.value!;
+        return ref entry.Value!;
     }
 
     /// <inheritdoc cref="IEnumerable{T}.GetEnumerator"/>
@@ -264,7 +264,7 @@ internal class Dictionary2<TKey, TValue> : IDictionarySlim<TKey, TValue>
     public Enumerator GetEnumerator() => new(this);
 
     /// <summary>
-    /// Enumerator for <see cref="DictionarySlim{TKey,TValue}"/>.
+    /// Enumerator for <see cref="Dictionary2{TKey,TValue}"/>.
     /// </summary>
     public ref struct Enumerator
     {
@@ -315,7 +315,7 @@ internal class Dictionary2<TKey, TValue> : IDictionarySlim<TKey, TValue>
             // For the same reason, we also removed the KeyValuePair<TKey, TValue> field here, and instead
             // rely on the properties lazily accessing the target instances directly from the current entry
             // pointed at by the index property (adjusted backwards to account for the increment here).
-            while (entries[this.index++].next < -1)
+            while (entries[this.index++].Next < -1)
             {
             }
 
@@ -328,7 +328,7 @@ internal class Dictionary2<TKey, TValue> : IDictionarySlim<TKey, TValue>
         public TKey Key
         {
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            get => this.entries[this.index - 1].key;
+            get => this.entries[this.index - 1].Key;
         }
 
         /// <summary>
@@ -337,7 +337,7 @@ internal class Dictionary2<TKey, TValue> : IDictionarySlim<TKey, TValue>
         public TValue Value
         {
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            get => this.entries[this.index - 1].value!;
+            get => this.entries[this.index - 1].Value!;
         }
     }
 
@@ -362,16 +362,16 @@ internal class Dictionary2<TKey, TValue> : IDictionarySlim<TKey, TValue>
             }
 
             entry = ref entries[i];
-            if (entry.hashCode == hashCode && entry.key.Equals(key))
+            if (entry.HashCode == hashCode && entry.Key.Equals(key))
             {
                 goto ReturnFound;
             }
 
-            i = entry.next;
+            i = entry.Next;
         } while (true);
 
         ReturnFound:
-        ref TValue value = ref entry.value!;
+        ref TValue value = ref entry.Value!;
 
         Return:
         return ref value;
@@ -417,11 +417,11 @@ internal class Dictionary2<TKey, TValue> : IDictionarySlim<TKey, TValue>
 
         for (int i = 0; i < count; i++)
         {
-            if (entries[i].next >= -1)
+            if (entries[i].Next >= -1)
             {
-                ref int bucket = ref GetBucket(entries[i].hashCode);
+                ref int bucket = ref GetBucket(entries[i].HashCode);
 
-                entries[i].next = bucket - 1;
+                entries[i].Next = bucket - 1;
                 bucket = i + 1;
             }
         }
@@ -448,138 +448,26 @@ internal class Dictionary2<TKey, TValue> : IDictionarySlim<TKey, TValue>
     private struct Entry
     {
         /// <summary>
-        /// The cached hashcode for <see cref="key"/>;
+        /// The cached hashcode for <see cref="Key"/>;
         /// </summary>
-        public uint hashCode;
+        public uint HashCode;
 
         /// <summary>
         /// 0-based index of next entry in chain: -1 means end of chain
         /// also encodes whether this entry this.itself_ is part of the free list by changing sign and subtracting 3,
         /// so -2 means end of free list, -3 means index 0 but on free list, -4 means index 1 but on free list, etc.
         /// </summary>
-        public int next;
+        public int Next;
 
         /// <summary>
         /// The key for the value in the current node.
         /// </summary>
-        public TKey key;
+        public TKey Key;
 
         /// <summary>
         /// The value in the current node, if present.
         /// </summary>
-        public TValue? value;
-    }
-
-    /// <summary>
-    /// A helper class for <see cref="Dictionary2{TKey,TValue}"/>.
-    /// </summary>
-    internal static partial class HashHelpers
-    {
-        /// <summary>
-        /// Maximum prime smaller than the maximum array length.
-        /// </summary>
-        private const int MaxPrimeArrayLength = 0x7FFFFFC3;
-
-        private const int HashPrime = 101;
-
-        /// <summary>
-        /// Table of prime numbers to use as hash table sizes.
-        /// </summary>
-        private static readonly int[] primes =
-        {
-            3, 7, 11, 17, 23, 29, 37, 47, 59, 71, 89, 107, 131, 163, 197, 239, 293, 353, 431, 521, 631, 761, 919,
-            1103, 1327, 1597, 1931, 2333, 2801, 3371, 4049, 4861, 5839, 7013, 8419, 10103, 12143, 14591,
-            17519, 21023, 25229, 30293, 36353, 43627, 52361, 62851, 75431, 90523, 108631, 130363, 156437,
-            187751, 225307, 270371, 324449, 389357, 467237, 560689, 672827, 807403, 968897, 1162687, 1395263,
-            1674319, 2009191, 2411033, 2893249, 3471899, 4166287, 4999559, 5999471, 7199369
-        };
-
-        /// <summary>
-        /// Checks whether a value is a prime.
-        /// </summary>
-        /// <param name="candidate">The value to check.</param>
-        /// <returns>Whether or not <paramref name="candidate"/> is a prime.</returns>
-        private static bool IsPrime(int candidate)
-        {
-            if ((candidate & 1) != 0)
-            {
-                int limit = (int)Math.Sqrt(candidate);
-
-                for (int divisor = 3; divisor <= limit; divisor += 2)
-                {
-                    if ((candidate % divisor) == 0)
-                    {
-                        return false;
-                    }
-                }
-
-                return true;
-            }
-
-            return candidate == 2;
-        }
-
-        /// <summary>
-        /// Gets the smallest prime bigger than a specified value.
-        /// </summary>
-        /// <param name="min">The target minimum value.</param>
-        /// <returns>The new prime that was found.</returns>
-        public static int GetPrime(int min)
-        {
-            foreach (int prime in primes)
-            {
-                if (prime >= min)
-                {
-                    return prime;
-                }
-            }
-
-            for (int i = min | 1; i < int.MaxValue; i += 2)
-            {
-                if (IsPrime(i) && ((i - 1) % HashPrime != 0))
-                {
-                    return i;
-                }
-            }
-
-            return min;
-        }
-
-        /// <summary>
-        /// Returns size of hashtable to grow to.
-        /// </summary>
-        /// <param name="oldSize">The previous table size.</param>
-        /// <returns>The expanded table size.</returns>
-        public static int ExpandPrime(int oldSize)
-        {
-            int newSize = 2 * oldSize;
-
-            if ((uint)newSize > MaxPrimeArrayLength && MaxPrimeArrayLength > oldSize)
-            {
-                return MaxPrimeArrayLength;
-            }
-
-            return GetPrime(newSize);
-        }
-
-        /// <summary>
-        /// Returns approximate reciprocal of the divisor: ceil(2**64 / divisor).
-        /// </summary>
-        /// <remarks>This should only be used on 64-bit.</remarks>
-        public static ulong GetFastModMultiplier(uint divisor)
-        {
-            return ulong.MaxValue / divisor + 1;
-        }
-
-        /// <summary>
-        /// Performs a mod operation using the multiplier pre-computed with <see cref="GetFastModMultiplier"/>.
-        /// </summary>
-        /// <remarks>This should only be used on 64-bit.</remarks>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static uint FastMod(uint value, uint divisor, ulong multiplier)
-        {
-            return (uint)(((((multiplier * value) >> 32) + 1) * divisor) >> 32);
-        }
+        public TValue? Value;
     }
 
     /// <summary>