… CommunityToolkit.Mvvm/Messaging/Internals/Microsoft.Collections.Extensions/DictionarySlim{TKey,TValue}.cs
… CommunityToolkit.Mvvm/Messaging/Internals/Microsoft.Collections.Extensions/HashHelpers.cs
… CommunityToolkit.Mvvm/Messaging/Internals/System/Collections.Generic/HashHelpers.cs
… CommunityToolkit.Mvvm/Messaging/Internals/System/Collections.Generic/IDictionary2.cs
… CommunityToolkit.Mvvm/Messaging/Internals/System/Collections.Generic/IDictionary2{TKey,TValue}.cs
… CommunityToolkit.Mvvm/Messaging/Internals/System/Collections.Generic/IDictionary2{TKey}.cs
… CommunityToolkit.Mvvm/Messaging/Internals/System/Runtime.CompilerServices/ConditionalWeakTable2{TKey,TValue}.Proxy.cs
… CommunityToolkit.Mvvm/Messaging/Internals/System/Runtime.CompilerServices/ConditionalWeakTable2{TKey,TValue}.ZeroAlloc.cs
… CommunityToolkit.Mvvm/Messaging/Internals/System/Runtime.CompilerServices/ConditionalWeakTable2{TKey,TValue}.cs
… CommunityToolkit.Mvvm/Messaging/Internals/System/Runtime.CompilerServices/ConditionalWeakTableExtensions.cs
… CommunityToolkit.Mvvm/Messaging/StrongReferenceMessenger.cs
… CommunityToolkit.Mvvm/Messaging/WeakReferenceMessenger.cs
… tests/CommunityToolkit.Mvvm.UnitTests/Internals/Test_ConditionalWeakTable2.cs
… tests/CommunityToolkit.Mvvm.UnitTests/Internals/Test_Dictionary2.cs
… tests/CommunityToolkit.Mvvm.UnitTests/Test_Messenger.cs