**Closes #40**

This PR includes the following changes:
- `WeakReferenceMessenger.Send` is now zero-alloc 🚀
- `WeakReferenceMessenger.Send` is now also ~2.4x faster 🚀
- Better collision resistance for registered handlers
- Improved lookup speed when using custom tokens
- Remove UB due to `Unsafe.As<T>(object)` used to alias delegate types (see dotnet/runtime#57495)

## Benchmarks

|                                 Info |     Mean |     Error |    StdDev |   Gen 0 | Allocated |
|------------------------------------- |---------:|----------:|----------:|--------:|----------:|
|                                 main | 17.41 ms | 0.2930 ms | 0.2600 ms | 31.2500 |    188 KB |
|           ConditionalWeakTable2 fork (https://github.com/CommunityToolkit/dotnet/pull/44/commits/dabca2ab242cbe1572c5f9f73f96da76b7bcd632) | 17.84 ms | 0.0820 ms | 0.0690 ms | 31.2500 |    188 KB |
|                ref struct enumerator (https://github.com/CommunityToolkit/dotnet/pull/44/commits/9b86a8844f3dd7a784902f01027c1422c00ca04d) | 15.32 ms | 0.0730 ms | 0.0570 ms |       - |         - |
|       No active enumerators tracking (https://github.com/CommunityToolkit/dotnet/pull/44/commits/b1dc97bb004f5a322eb04b00d2bfa3c5c9efb6ba) | 15.25 ms | 0.0260 ms | 0.0200 ms |       - |         - |
|          Monitor.Enter loop hoisting (https://github.com/CommunityToolkit/dotnet/pull/44/commits/e92c5c28c81fdc4c4c96bf0df691355833c7a666) | 11.33 ms | 0.2090 ms | 0.1960 ms |       - |         - |
|                     Dictionary2 fork (https://github.com/CommunityToolkit/dotnet/pull/44/commits/9be01faee04e89eba7df26f0828aa74c45b490a1) | 11.52 ms | 0.0370 ms | 0.0290 ms |       - |         - |
|          typeof(Unit) specialization (https://github.com/CommunityToolkit/dotnet/pull/44/commits/fc538c385c1f46231991bd0c3780456759a4c884) | 9.839 ms | 0.1863 ms | 0.1743 ms |       - |         - |
|          Remove Unsafe.As<T>(object) (https://github.com/CommunityToolkit/dotnet/pull/44/commits/9dd8ea9ea0936d8edcb40cb3292080d773fef54c) | 9.830 ms | 0.1930 ms | 0.1982 ms |       - |         - |
| No bounds checks, strength reduction (https://github.com/CommunityToolkit/dotnet/pull/44/commits/4408379bea394e50c9714c60fd32c395aa01ed56) | 9.309 ms | 0.1709 ms | 0.1515 ms |       - |         - |
|         Remove KeyValuePair<,> usage | 7.980 ms | 0.1586 ms | 0.2006 ms |       - |         - |
|     Reduce enumerator field accesses (https://github.com/CommunityToolkit/dotnet/pull/44/commits/3f20dce74b84d5faf597936effa979dff4bb12bf) | 7.348 ms | 0.1494 ms | 0.1834 ms |       - |         - |
|       Remove array covariance checks (https://github.com/CommunityToolkit/dotnet/pull/44/commits/15bd652a572fbf7735ba53e62945473d03fa6add) | **7.246 ms** | 0.1056 ms | 0.0988 ms |       **-** |         **-** |